Fix weird decimal casting behaviour.
Usage.md helpfully explains:

> Note that I use
`'type' => 'DECIMAL(10,7)'` so that sorting is done numerically, instead of alphabetically.

However, for reasons beyond me, it seems that when casting these distances as a [DECIMAL](https://dev.mysql.com/doc/refman/8.0/en/fixed-point-types.html) type using a *scale* greater than 2 they do not order correctly. This is causing the OrderBy tests to fail.

Replacing with DECIMAL(10,2), DECIMAL(10) and DECIMAL() all cause the expected behaviour (treated as numeric and ordered accordingly).

Interestingly enough, on MySQL 8 / WP6 even when the `type` parameter is omitted and the value is cast as a CHAR it orders the test numbers correctly.

This change uses DECIMAL(10) and tests pass as expected.